requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
vosk>=0.3.45
pydub>=0.25.1
ffmpeg-python>=0.2.0
numpy>=1.24.0
PyNaCl>=1.5.0